requirements.txt
Django
django_crispy_forms
python_dateutil
Requests
